Adds more thoughts on DB design.
The design of the database shouldnt' change once I start working
with the app. I'd like to have the design well thought-out before
any work gets done so that I don't have to mess around with it
again. It's a real pain to do, and if I switch to a different
schema halfway through testing there will be a bunch of errors
due to a mix of different schemas.