Adds final thoughts on DB schema design.
A single collection of user IDs with associated information should work
perfectly; I see no need for three separate collections for user info,
plans, and feedback. The session can grab the information out of the
database when it's initialized, modify it in memory as the user makes
changes, then save to the database again. Or something like that.