Cleans up demo CSS/HTML, adds some real content.
"Real content" is really just the names of different sections of the
site. I'm still working off of the rough sketches I did in my notebook
and in Mr. Klick's room. The idea is to get a scaffolding done here on
the UI side before I write a ton of backend code. I'm still figuring out
exactly what pages I'll need to show, and what functionality they'll
have to contain.