Adds the logic for getting DB variables.
There are different database configurations for server deploys and local
testing. This adds a function `GetDBData()` to retrieve these variables
no matter where the code is run. The code that makes up this function
used to be part of `routes.js`, but it makes more sense to put it in
this module.