Adds functional login and logout methods.
A user can now log in and out of the website from the navbar. I have yet
to write the code for logging in from a specific "login" page — that is
on its way. The authentication is verified on the client side very
nicely, to make sure that they are submitting the right variables.