Basic structure of signup page.
I've been working on this all afternoon, although it may not seem like
a ton of result. I figured out how to do a nice structure for the
different fields of the forms. There is also validation on these fields.
I have the structure figured out to a point where I should be able to
create a Jade mixin for each field (except for the password one
— there's some weird stuff when it comes to having two different fields
in the same `control-group`).