Signup now sends email through to login page.
Before, the email wasn't properly being sent through. The reason is that
there are two elements with the ID #email: the one in the login form and
the one in the signup form. I should probably either change their names
or just not have them both in the same place.