Improves signup experience.
Now, when the user clicks "Sign Me Up", the button changes to say
"Creating your account ...", giving them feedback that the request is
underway. If there's an error, the website shoots out an alert saying
so. In the future, I'd like to change the alert dialog to instead be a
little message on the site itself.