Full dummy dashboard.
I had a very hard time coming up with the layout for this but I think it
looks pretty cool. I'd like to know what Mr. Klick thinks about it, but
it seems *right* to me. It's simple, clean, and tells the user what they
need to know. Also, right now it doesn't work — none of it is automated.
That being said, it's designed in a way that I can easily plug in the
correct values once the calendar generation is done. That's what I need
to work on next, and I'd like to do it with Mr. Klick. Ideally, I'll
just hack on it in his classroom and have him available for questions.