Improves layout text and color.
I've decided that the 'quick history' in the lower right should have the
same 'well' look as the rest of the items. The page is now symmetrical
and looks quite nice. I've also added a third row of history and a link
to see more statistics. Depending on how tall the logging form is, I'll
change the number of past weeks.